Add Twitter gem and client factory
The idea here is that upon booting the app, we have what we need to make
a client, but we don't actually generate it until we need to use it.
Creating a client is as easy as:

```ruby
client = TwitterClient.generate
```

And should a mistake be made where you didn't setup your ENV properly,
then you'll get the auth errors from the Twitter gem.

Closes #3